+++ services/compute/mgmt/2021-07-01/compute/CHANGELOG.md
@@ -1,2 +1,27 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Structs
+
+1. RestorePointProvisioningDetails
+
+#### Removed Struct Fields
+
+1. RestorePointProperties.ProvisioningDetails
+
+### Signature Changes
+
+#### Struct Fields
+
+1. OrchestrationServiceStateInput.ServiceName changed type from OrchestrationServiceNames to *string
+
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. RestorePointProperties.TimeCreated