+++ services/compute/mgmt/2021-12-01/compute/CHANGELOG.md
@@ -1,2 +1,23 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Struct Fields
+
+1. OrchestrationServiceStateInput.ServiceName changed type from OrchestrationServiceNames to *string
+1. SharingProfile.CommunityGalleryInfo changed type from interface{} to *CommunityGalleryInfo
+
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. UserArtifactSettings
+
+#### New Struct Fields
+
+1. GalleryApplicationVersionPublishingProfile.AdvancedSettings
+1. GalleryApplicationVersionPublishingProfile.Settings